sdl tetris

+++ stodin-lib/__stodin_types.h
@@ -3,6 +3,6 @@
 
 #include <cstdint>
 
-typedef uint32_t __stodin_bool;
+typedef uint8_t __stodin_bool;
 
 #endif // __STODIN_TYPES_H_INCLUDED
